chore(live-e2e): fix lint and split lock module for size gate
